# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/amadeus/experiences/activities": {
      "get": {
        "operationId": "AmadeusExperiencesController_getActivities",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 41.397158,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2.160873,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Tours & Activities by location",
        "tags": [
          "Amadeus — Destination Experiences"
        ]
      }
    },
    "/api/v1/amadeus/experiences/activities/by-square": {
      "get": {
        "operationId": "AmadeusExperiencesController_getActivitiesBySquare",
        "parameters": [
          {
            "name": "north",
            "required": true,
            "in": "query",
            "schema": {
              "example": 41.397158,
              "type": "number"
            }
          },
          {
            "name": "west",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2.160873,
              "type": "number"
            }
          },
          {
            "name": "south",
            "required": true,
            "in": "query",
            "schema": {
              "example": 41.394582,
              "type": "number"
            }
          },
          {
            "name": "east",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2.177181,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Tours & Activities within a geographic square",
        "tags": [
          "Amadeus — Destination Experiences"
        ]
      }
    },
    "/api/v1/amadeus/experiences/activities/{activityId}": {
      "get": {
        "operationId": "AmadeusExperiencesController_getActivityById",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a specific tour or activity by ID",
        "tags": [
          "Amadeus — Destination Experiences"
        ]
      }
    },
    "/api/v1/amadeus/experiences/cities": {
      "get": {
        "operationId": "AmadeusExperiencesController_citySearch",
        "parameters": [
          {
            "name": "countryCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "FR",
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PARIS",
              "type": "string"
            }
          },
          {
            "name": "max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {
              "example": "AIRPORTS",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "City search",
        "tags": [
          "Amadeus — Destination Experiences"
        ]
      }
    },
...
    "/api/v1/amadeus/hotels/by-city": {
      "get": {
        "operationId": "AmadeusHotelsController_hotelsByCity",
        "parameters": [
          {
            "name": "cityCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "DEL",
              "type": "string"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusUnit",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "KM",
                "MILE"
              ],
              "type": "string"
            }
          },
          {
            "name": "amenities",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratings",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotelSource",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "BEDBANK",
                "DIRECTCHAIN",
                "ALL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List hotels by city code",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/by-geocode": {
      "get": {
        "operationId": "AmadeusHotelsController_hotelsByGeocode",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 41.397158,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2.160873,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusUnit",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "KM",
                "MILE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List hotels by geographic coordinates",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/by-hotels": {
      "get": {
        "operationId": "AmadeusHotelsController_hotelsByIds",
        "parameters": [
          {
            "name": "hotelIds",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MCLONGHM",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List hotels by hotel IDs",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/offers": {
      "get": {
        "operationId": "AmadeusHotelsController_searchOffers",
        "parameters": [
          {
            "name": "hotelIds",
            "required": true,
            "in": "query",
            "description": "Comma-separated Amadeus hotel IDs",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adults",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "checkInDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkOutDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomQuantity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceRange",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currency",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bestRateOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search hotel offers",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/offers/{offerId}": {
      "get": {
        "operationId": "AmadeusHotelsController_getOffer",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get specific hotel offer details",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/booking/v1": {
      "post": {
        "operationId": "AmadeusHotelsController_bookV1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create hotel booking (v1)",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/booking/v2": {
      "post": {
        "operationId": "AmadeusHotelsController_bookV2",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create hotel booking (v2 — hotel orders)",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/ratings": {
      "get": {
        "operationId": "AmadeusHotelsController_ratings",
        "parameters": [
          {
            "name": "hotelIds",
            "required": true,
            "in": "query",
            "schema": {
              "example": "ELONMFS,ADNYCCTB",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Hotel sentiments / ratings",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
    "/api/v1/amadeus/hotels/autocomplete": {
      "get": {
        "operationId": "AmadeusHotelsController_autocomplete",
        "parameters": [
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PARI",
              "type": "string"
            }
          },
          {
            "name": "subType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "HOTEL_LEISURE",
                "HOTEL_GDS"
              ],
              "type": "string"
            }
          },
          {
            "name": "countryCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Hotel name autocomplete",
        "tags": [
          "Amadeus — Hotels"
        ]
      }
    },
...
  },
  "info": {
    "title": "✈️ Flight Booking API",
    "description": "\n      ## Multi-Tenant Flight Booking Platform API\n      \n      A scalable flight booking backend with:\n      - **Multi-tenancy** (B2B & B2C)\n      - **Multi-currency** support (14 currencies)\n      - **Dual payment** processing (Stripe + Paystack)\n      - **Real-time** seat availability\n      - **Admin analytics** and reporting\n      - **Global Travel Integration** (Amadeus Search, Price, Booking)\n      \n      ### Amadeus API Modules\n      This platform integrates the following Amadeus modules:\n      - **Flights**: Live search, pricing, and orders (31 endpoints)\n      - **Hotels**: Search, offers, and bookings (10 endpoints)\n      - **Experiences**: Tours, activities, and city search (4 endpoints)\n      - **Transfers**: Cars and transfers search/booking (3 endpoints)\n      - **Insights**: Market trends and traffic analytics (3 endpoints)\n      - **Itinerary**: Travel predictions and planning (1 endpoint)\n      \n      ### Authentication\n      All protected endpoints require a Bearer JWT token.\n      Use `/auth/login` to obtain tokens.\n      \n      ### Rate Limiting\n      API requests are rate-limited to 100 requests per minute per IP.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}